requirements.txt
blinker==1.7.0
boto3==1.34.70
botocore==1.34.70
click==8.1.7
Flask==3.0.2
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
pillow==10.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.18.1
